rename restaurant
